+++ src/components/ProductCatalogComponent/SidebarComponent.vue
